+++ js/main.js
@@ -1,5 +1,5 @@
 import './similar-place.js';
-import './form.js';
-import './map.js';
+//import './form.js';
+//import './map.js';
 import './server.js';
 //import './filter.js';
